fix integer overflow issue in isomedia/box_code_base.c:413

+++ include/gpac/internal/isomedia_dev.h
@@ -1189,7 +1189,7 @@ typedef struct
 	u32 r_currentEntryIndex;
 	u32 r_FirstSampleInEntry;
 
-	s32 max_cts_delta;
+	u64 max_cts_delta;
 	//u32 sample_num_max_cts_delta;
 } GF_CompositionOffsetBox;
